AutoPLM-oaz/AutoPLM-recuperacao-autoplm/requirements.txt
email-validator>=2.3.0
flask>=3.1.2
flask-sqlalchemy>=3.1.1
flask-wtf>=1.2.2
openai>=1.107.2
pandas>=2.2.3
pillow>=12.0.0
psycopg2-binary>=2.9.10
openpyxl>=3.1.5
pymupdf>=1.26.5
pytesseract>=0.3.10
pypdf2>=3.0.1
python-dotenv>=1.1.1
replit-object-storage>=1.0.2; python_version < "3.13"
werkzeug>=3.1.3
wtforms>=3.2.1
httpx>=0.27.0
tenacity>=8.0.0
